Move PathValidator adjacent to other filesystem bits
